+++ src/main/java/com/quartzshard/aasb/api/alchemy/IAlchemicalFlow.java
@@ -4,21 +4,21 @@ public interface IAlchemicalFlow<T> {
 	/**
 	 * Checks if the caller flows into the arg. Order matters!
 	 * @param to The Shape we're checking flow to.
-	 * @return 
+	 * @return if this follows flow
 	 */
 	public boolean flows(T to);
 	
 	/**
 	 * Checks if the caller is perpendicular to the arg. Order matters!
 	 * @param to The Shape we're checking against.
-	 * @return
+	 * @return if this is considered "perpendicular" (border between flowing & not flowing)
 	 */
 	public boolean perpendicular(T to);
 	
 	/**
 	 * Checks if flow is violated when travelling from caller to arg. Order matters!
 	 * @param to The Shape we're checking against.
-	 * @return
+	 * @return if this violates flow
 	 */
 	public boolean violates(T to);
 }
